+++ frontend/src/components/Channel/Channel.jsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import useFetch from '@/hooks/useFetch';
 import useSocket from '@/hooks/useSocket';
 
-import Video from '@/components/Video';
+import HLSVideo from '@/components/HLS/HLSVideo';
 import Chat from '@/components/Chat';
 import Box from '@/components/Common/Box';
 import ChannelDetail from './ChannelDetail';
@@ -28,7 +28,7 @@ export default function Channel() {
         <Box flex={1} width="100%" height="100%" alignItems="flex-start">
             <Box flexDirection="column" height="100%" flex={3}>
                 <Box width="100%" flex={5}>
-                    <Video streamKey={data.streamKey} />
+                    <HLSVideo streamKey={data.streamKey} />
                 </Box>
 
                 <Box
