+++ frontend/src/components/HLS/HLSVideo/HLSVideo.stories.jsx
@@ -1,0 +1,11 @@
+import HLSVideo from './HLSVideo';
+
+export default {
+    title: 'HLSVideo',
+    component: HLSVideo,
+};
+
+export const Default = HLSVideo.bind();
+Default.args = {
+    secretKey: 'output',
+};
