+++ frontend/src/components/HLS/HLSVideo/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './HLSVideo';
