+++ frontend/src/components/HLS/HLSVideoOverlay/index.jsx
@@ -1,0 +1,1 @@
+export {default} from './HLSVideoOverlay';
