--- frontend/src/components/Video/Video.stories.jsx
@@ -1,11 +1,0 @@
-import Video from './Video';
-
-export default {
-    title: 'Video',
-    component: Video,
-};
-
-export const Default = Video.bind();
-Default.args = {
-    secretKey: 'output',
-};
