--- frontend/src/components/Video/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './Video';
