--- frontend/src/components/VideoOverlay/index.jsx
@@ -1,1 +1,0 @@
-export {default} from './VideoOverlay';
